Long

+++ integrator.h
@@ -1,2 +1,2 @@
-int velverlet_ndim_npart(const double dt, const double forceConst, const int nBodies, const int spatialDim, const double *masses,
-double *coord, double *vel, double *force, double **f_o, void (*F)(const double *, const double *, const double, const int, double *));
+int velverlet_ndim_npart(const long double dt, const long double forceConst, const int nBodies, const int spatialDim, const long double *masses,
+                         long double *coord, long double *vel, long double *force, long double **f_o, void (*F)(const long double *, const long double *, const long double, const int, long double *));
